+++ src/src.cmake
@@ -255,8 +255,6 @@ if (SOLOUD_BACKEND_MINIAUDIO)
 			)
 
 	add_definitions(-DWITH_MINIAUDIO)
-	# Remove PulseAudio since it can cause stutters and glitches
-	add_definitions(-DMA_NO_PULSEAUDIO)
 	set (BACKENDS_SOURCES
 			${BACKENDS_SOURCES}
 			${BACKENDS_PATH}/miniaudio/miniaudio.h